Retarget from .NET 4.5 to 4.8, updated protocols and rework in Sample project

+++ source/ChromeDevTools/Protocol/Chrome/IndexedDB/DatabaseWithObjectStores.cs
@@ -5,21 +5,26 @@ using System.Collections.Generic;
 namespace MasterDevs.ChromeDevTools.Protocol.Chrome.IndexedDB
 {
 	/// <summary>
-	/// Database with an array of object stores.
+	///Database with an array of object stores.
+
 	/// </summary>
 	[SupportedBy("Chrome")]
 	public class DatabaseWithObjectStores
 	{
 		/// <summary>
-		/// Gets or sets Database name.
+		/// Gets or setsDatabase name.
+
 		/// </summary>
 		public string Name { get; set; }
 		/// <summary>
-		/// Gets or sets Database version.
+		/// Gets or setsDatabase version (type is not 'integer', as the standard
+		/// requires the version number to be 'unsigned long long')
+
 		/// </summary>
-		public long Version { get; set; }
+		public double Version { get; set; }
 		/// <summary>
-		/// Gets or sets Object stores in this database.
+		/// Gets or setsObject stores in this database.
+
 		/// </summary>
 		public ObjectStore[] ObjectStores { get; set; }
 	}
